backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
motor==3.3.2
pydantic==2.6.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
pytest==8.0.0
httpx==0.26.0
celery==5.3.6
redis==5.0.1
gunicorn==21.2.0
email-validator==2.1.0.post1 